Assert allocator-type is known on OOME
We should always be using a known allocator-type. So instead of being
conditional about this, assert it.

This change also simplifies updation of alloc entrypoints for CC
collector. Earlier we used to choose TLAB allocator on x86_64 targets
to avoid a read-barrier in allocation path. Now that it's not there, we
can simplify to always use RegionTLAB allocator for CC.

Test: art/test/testrunner/testrunner.py -t 004-ThreadStress
Bug: 154100060
Change-Id: Ic692794c3dab80a67c2c4b3274adf00f879ad21b